Represented the queries as tree in the new queries browser (#152)
Represented the queries with indentation as trees. This is done simulating a tree with a component list with spaces in the layout of each of its presenters..
Now a query with children can be deleted. This will delete all children too.
Now a query with children can add one more child.
Cleaned the code and added more tests.