Speed up Models browser
The model browser can be really slow if we have big models in the image. This is due to the fact that propagating is filtering the stubs. But this means that #canPropagate it always filtering the stubs without using any cache (it would be hard to cache this). 

I propose to not filter in #canPropagate because we should almost never have a model without non stub entities and this will make the tools way smoother. We'll filter only on the propagate.

Fixes #976