Fix the display of tags in the disctribution map settings
Currently to have some tags shown in the distribution map settings, we need the tag browser open to be a provider of tags which is not really user friendly. 

I propose another way to get the tags. Ideally we should have entities of only one model for the distribution map (except if we play by hand in a playground and propagate this..). So I now ask a random entity of the model and gets the tags from its model.

Fixes #985