Architectural: display name of parents for entities in tags
When we open the architectural on tags it might be hard to distinguish the entities. For example I have a cas where I had 15 methods names #write tagged with the same tag. 

With Imen we propose to display also the name of the parent entity in the label to recognize better the entities when their parent is a tag