Move model detection code
In the model importer we have some code to be able to find which model to use for a specific file to import. I'm moving this code from the interface to the importer so that I can reuse it in my softwase heritage code without depending on the MooseIDE interface.

I also updated the layout of the import dialog to ask for the file first since we auto fill the model type.